+++ Courses/CoursesApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct CoursesApp: App {
     var body: some Scene {
         WindowGroup {
-            PersonView()
+            TabBar()
         }
     }
 }
