+++ Courses/Navigation/ContentView.swift
@@ -1,0 +1,45 @@
+//
+//  ContentView.swift
+//  Courses
+//
+//  Created by Dmytro Churakov on 02.10.2022.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    @AppStorage("selectedTab") var selectedTab: Tab = .home
+    var body: some View {
+        ZStack(alignment: .bottom) {
+            
+            Group {
+                switch selectedTab {
+                case .home:
+                    HomeView()
+                case .explore:
+                    PersonView()
+                case .notifications:
+                    HomeView()
+                case .account:
+                    PersonView()
+                }
+            }
+            .frame(maxWidth: .infinity, maxHeight: .infinity)
+            
+            TabBar()
+        }
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        Group {
+            ContentView()
+                .previewDevice("iPhone 13")
+            ContentView()
+                .preferredColorScheme(.dark)
+                .previewDevice("iPhone 13")
+        }
+        
+    }
+}
